// reinvest-web-app/src/views/account-management/enums/flow.ts
export enum FlowIdentifiers {
  INVESTMENT_HISTORY = 'INVESTMENT_HISTORY',
  RECURRING_INVESTMENTS = 'RECURRING_INVESTMENTS',
  DIVIDEND_REINVESTING = 'DIVIDEND_REINVESTING',
  WITHDRAW_FUNDS = 'WITHDRAW_FUNDS',
  BANK_ACCOUNT = 'BANK_ACCOUNT',
  ACCOUNT_ACTIVITY = 'ACCOUNT_ACTIVITY',

  EMAIL_ADDRESS = 'EMAIL_ADDRESS',
  PHONE_NUMBER = 'PHONE_NUMBER',
  CHANGE_PASSWORD = 'CHANGE_PASSWORD',

  NAME = 'NAME',
  UPDATE_PROFILE_PICTURE = 'UPDATE_PROFILE_PICTURE',
  UPDATE_ADDRESS = 'UPDATE_ADDRESS',
}